requirements.txt
# HISM: Highlight-Informed Saliency Model
# Requirements file

# Core ML framework
torch>=2.0.0
torchvision>=0.15.0

# Data processing
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.0.0

# Configuration
pyyaml>=6.0
